Use ints instead of doubles for storing trial start capacity

+++ PowerGremlin/PGPowerDetailsTableViewController.h
@@ -10,7 +10,7 @@
 
 @interface PGPowerDetailsTableViewController : UITableViewController
 
-@property(nonatomic) double trialStartCapacity;
+@property(nonatomic) int trialStartCapacity;
 @property(nonatomic) double trialStartTime;
 
 @end
